Significant refactor of Elasticsearch ETL scripts
